requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
user_agent==0.1.10
waitress==3.0.0
Werkzeug==3.0.2
